requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
wheel==0.45.1
openai==1.68.2
anthropic==0.47.1
google-auth==2.38.0
pyjwt==2.10.1
requests==2.32.3
litellm==1.72.2
pydantic-settings==2.8.1
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
dependency-injector==4.46.0
asyncpg==0.30.0
SQLAlchemy[asyncio]==2.0.40
python-multipart==0.0.20
cryptography==44.0.2
google-cloud-storage==3.1.0
duckduckgo-search==8.0.2
beautifulsoup4==4.13.4
slowapi==0.1.9